ads_service/requirements.txt
google-ads
uvicorn
fastapi
psycopg2-binary
